moved files again and changed .csproj
added BannerlordPath variable to targets, which specifies the path to games directory. It only needs to be changed in one place, but keeps references consistent across other contributors. However, every time someone changes it, they should take care not to commit the change, which makes this a temporary solution at best