Fixed the logic
the main problem was in CrippleHorse method. Lists, Dictionaries, and other IEnumerables keep stuff inside as references, instead of copying values. So, what ended up happening, is that same exact queue was being added to the dictionary, and then emptied. So, the item in the dictionary was emptied again. That means that the debuffs were not being removed.
By copying the queue into a new one, then adding the new one fixes the problem. Bear in mind, I only tested it once at 1000 battle size (couldnt debug, was too slow)